Compiles and all modes seem to work and produce accurate results.
However asynchronous works only with compressed data option. Need to
generalize it.

+++ src/compression.cpp
@@ -16,7 +16,7 @@ unsigned long maxCompressedDataSize(const unsigned int numFloats)
         return 0;
 
     // Calculate the maximum buffer size needed to hold the compressed data
-    const unsigned int inputSize = numFloats * sizeof(float);
+    const unsigned int inputSize = numFloats * sizeof(double);
     strm.avail_in = inputSize;
     const auto maxOutputSize = deflateBound(&strm, inputSize);
     //std::cout << "maxSize = " << maxOutputSize << " bytes = " << maxOutputSize / 1024 << " KiB" << std::endl;
@@ -27,7 +27,7 @@ unsigned long maxCompressedDataSize(const unsigned int numFloats)
     return maxOutputSize;
 }
 
-unsigned long compressData(const VectorXf &snpData, unsigned char *outputBuffer, unsigned long outputSize)
+unsigned long compressData(const VectorXd &snpData, unsigned char *outputBuffer, unsigned long outputSize)
 {
     // Initialise zlib
     z_stream strm;
@@ -40,9 +40,9 @@ unsigned long compressData(const VectorXf &snpData, unsigned char *outputBuffer,
         return 0;
 
     // Compress the data
-    const unsigned int inputSize = static_cast<unsigned int>(snpData.size()) * sizeof(float);
+    const unsigned int inputSize = static_cast<unsigned int>(snpData.size()) * sizeof(double);
     strm.avail_in = inputSize;
-    strm.next_in = reinterpret_cast<unsigned char *>(const_cast<float*>(&snpData[0]));
+    strm.next_in = reinterpret_cast<unsigned char *>(const_cast<double*>(&snpData[0]));
     strm.avail_out = static_cast<unsigned int>(outputSize);
     strm.next_out = outputBuffer;
 
@@ -93,15 +93,15 @@ unsigned long compressData(const VectorXf &snpData, unsigned char *outputBuffer,
     // Cleanup
     delete[] checkBuffer;
     (void) inflateEnd(&strm2);
-     */
+    */
 
     return compressedSize;
 }
 
 void extractData(unsigned char *compressedData,
-        unsigned int compressedDataSize,
-        unsigned char *outputBuffer,
-        unsigned int outputBufferSize)
+                 unsigned int compressedDataSize,
+                 unsigned char *outputBuffer,
+                 unsigned int outputBufferSize)
 {
     z_stream strm;
     strm.zalloc = nullptr;
