+++ src/data.hpp
@@ -86,8 +86,8 @@ public:
 
     // mmap related data
     int ppBedFd;
-    float *ppBedMap;
-    Map<MatrixXf> mappedZ;
+    double *ppBedMap;
+    Map<MatrixXd> mappedZ;
     PpBedIndex ppbedIndex;
 
     // Original data
@@ -133,6 +133,7 @@ public:
     void preprocessBedFile(const string &bedFile, const string &preprocessedBedFile, const string &preprocessedBedIndexFile, bool compress);
     void mapPreprocessBedFile(const string &preprocessedBedFile);
     void unmapPreprocessedBedFile();
+
     void mapCompressedPreprocessBedFile(const string &preprocessedBedFile, const string &indexFile);
     void unmapCompressedPreprocessedBedFile();
